changed batteryMilliAmps to float in h function in order to calculate correctly, it seems the algo was developed for current bigger than a certain amount of mA. Added some more debug info.

+++ src/SoCKalman.h
@@ -90,7 +90,7 @@ class SoCKalman
      * 
      * @param isBatteryLithium, batteryMilliAmps
      */
-    void h(int32_t batteryMilliAmps);
+    void h(float batteryMilliAmps);
 
     void diagonalMatrix(float value, float* result);
